removed commit_log/commit_log.rs
