+++ storage/Cargo.toml
@@ -43,3 +43,4 @@ serde_json = "1.0"
 slog-async = "2.6"
 slog-term = "2.6"
 criterion = "0.3"
+fs_extra = "1.2.0"
