+++ storage/src/persistent/mod.rs
@@ -19,7 +19,7 @@ use crate::backend::btree_map::BTreeMapBackend;
 use crate::backend::in_memory_backend::InMemoryBackend;
 use crate::backend::rocksdb_backend::RocksDBBackend;
 use crate::backend::sled_backend::SledBackend;
-pub use crate::commit_log::commit_log::{CommitLogError, CommitLogRef, CommitLogs, CommitLogWithSchema, Location};
+pub use crate::commit_log::{CommitLogError, CommitLogRef, CommitLogs, CommitLogWithSchema, Location};
 use crate::KeyValueStoreBackend;
 use crate::merkle_storage::MerkleStorage;
 use crate::persistent::sequence::Sequences;
